Fixed compilation due to zlib support being JVM only
