+++ jvm/src/main/scala/scodec/codecs/ZlibCodec.scala
